requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
cryptography
sphinx
shibuya
